refactor: add core::{structure,decoding}::types modules
