+++ src/core/indices.rs
@@ -36,8 +36,7 @@ use alloc::{boxed::Box, vec::Vec};
 
 use crate::{
     core::{
-        decoding::error::DecodingError,
-        reader::{types::FuncType, WasmReader},
+        decoding::error::DecodingError, reader::WasmReader, structure::types::FuncType,
         utils::ToUsizeExt,
     },
     ValType, ValidationError,
