+++ src/core/structure/mod.rs
@@ -1,1 +1,2 @@
 pub mod instructions;
+pub mod types;
