+++ src/execution/store/instances.rs
@@ -3,11 +3,9 @@ use alloc::{collections::btree_map::BTreeMap, string::String, vec, vec::Vec};
 use crate::{
     core::{
         indices::{DataIdx, ElemIdx, FuncIdx, GlobalIdx, IdxVec, MemIdx, TableIdx, TypeIdx},
-        reader::{
-            span::Span,
-            types::{FuncType, MemType, TableType},
-        },
+        reader::span::Span,
         sidetable::Sidetable,
+        structure::types::{FuncType, MemType, TableType},
         utils::ToUsizeExt,
     },
     linear_memory::LinearMemory,
