+++ src/execution/store/mod.rs
@@ -9,9 +9,9 @@ use crate::core::reader::span::Span;
 use crate::core::reader::types::data::{DataModeActive, DataSegment};
 use crate::core::reader::types::element::{ActiveElem, ElemItems, ElemMode, ElemType};
 use crate::core::reader::types::export::ExportDesc;
-use crate::core::reader::types::global::GlobalType;
-use crate::core::reader::types::{ExternType, FuncType, ImportSubTypeRelation, MemType, TableType};
+use crate::core::reader::types::ImportSubTypeRelation;
 use crate::core::reader::WasmReader;
+use crate::core::structure::types::{ExternType, FuncType, GlobalType, MemType, TableType};
 use crate::core::utils::ToUsizeExt;
 use crate::execution::interpreter_loop::{self, memory_init, table_init, InterpreterLoopOutcome};
 use crate::execution::value::{Ref, Value};
