+++ src/validation/code.rs
@@ -13,9 +13,9 @@ use crate::core::reader::span::Span;
 use crate::core::reader::types::element::ElemType;
 use crate::core::reader::types::global::Global;
 use crate::core::reader::types::memarg::MemArg;
-use crate::core::reader::types::{BlockType, FuncType, MemType, NumType, TableType, ValType};
 use crate::core::reader::WasmReader;
 use crate::core::sidetable::{Sidetable, SidetableEntry};
+use crate::core::structure::types::{BlockType, FuncType, MemType, NumType, TableType, ValType};
 use crate::core::utils::ToUsizeExt;
 use crate::validation_stack::{LabelInfo, ValidationStack};
 use crate::{RefType, ValidationError};
