+++ src/validation/globals.rs
@@ -3,8 +3,9 @@ use alloc::vec::Vec;
 
 use crate::core::indices::{FuncIdx, IdxVec, TypeIdx};
 use crate::core::reader::section_header::{SectionHeader, SectionTy};
-use crate::core::reader::types::global::{Global, GlobalType};
+use crate::core::reader::types::global::Global;
 use crate::core::reader::WasmReader;
+use crate::core::structure::types::GlobalType;
 use crate::read_constant_expression::read_constant_expression;
 use crate::validation_stack::ValidationStack;
 use crate::ValidationError;
