+++ src/validation/mod.rs
@@ -13,11 +13,13 @@ use crate::core::reader::span::Span;
 use crate::core::reader::types::data::DataSegment;
 use crate::core::reader::types::element::ElemType;
 use crate::core::reader::types::export::{Export, ExportDesc};
-use crate::core::reader::types::global::{Global, GlobalType};
+use crate::core::reader::types::global::Global;
 use crate::core::reader::types::import::{Import, ImportDesc};
-use crate::core::reader::types::{ExternType, FuncType, MemType, ResultType, TableType};
 use crate::core::reader::WasmReader;
 use crate::core::sidetable::Sidetable;
+use crate::core::structure::types::{
+    ExternType, FuncType, GlobalType, MemType, ResultType, TableType,
+};
 use crate::core::utils::ToUsizeExt;
 use crate::custom_section::CustomSection;
 use crate::ValidationError;
