+++ src/validation/read_constant_expression.rs
@@ -2,8 +2,8 @@ use alloc::vec::Vec;
 
 use crate::core::indices::{FuncIdx, IdxVec, TypeIdx};
 use crate::core::reader::span::Span;
-use crate::core::reader::types::global::GlobalType;
 use crate::core::reader::WasmReader;
+use crate::core::structure::types::GlobalType;
 use crate::core::utils::ToUsizeExt;
 use crate::{NumType, RefType, ValType, ValidationError};
