+++ src/validation/validation_stack.rs
@@ -5,10 +5,10 @@ use alloc::vec::Vec;
 
 use crate::{
     core::{
-        reader::types::{FuncType, ResultType},
+        structure::types::{FuncType, NumType, RefType, ResultType, ValType},
         utils::ToUsizeExt,
     },
-    NumType, RefType, ValType, ValidationError,
+    ValidationError,
 };
 
 #[derive(Debug, PartialEq, Eq)]
